+++ remote-script/__init__.py
@@ -1,0 +1,5 @@
+from .PushHackArrangement import PushHackArrangement
+
+
+def create_instance(c_instance):
+    return PushHackArrangement(c_instance)